fix: type error in next page export field
